Refactor database collection names for comments and shares, update API endpoints to use username for user identification, and enhance post fetching logic for improved clarity and consistency
