+++ src/components/posts.tsx
@@ -1,6 +1,6 @@
 "use client";
-import React, { ChangeEvent, Dispatch, ReactNode, RefObject, MouseEvent, use, useCallback, useEffect, useRef, useState } from 'react';
-import { Post, PostData, PostProps, formatNo, timeFormatter, updateLiveTime } from '../templates/PostProps';
+import React, { useCallback, useEffect, useState } from 'react';
+import { Post, PostData, formatNo, timeFormatter, updateLiveTime } from '../templates/PostProps';
 import { useRouter } from 'next/navigation';
 import {
   Popover,
@@ -8,7 +8,7 @@ import {
   PopoverTrigger,
 } from "@/components/ui/popover"
 import Link from 'next/link';
-import { AlignCenter, AlignLeft, AlignRight, Bold, ChartBarDecreasing, ChevronDown, ChevronLeft, ChevronRight, CircleAlert, CircleCheck, Clock4, Copy, Crop, Delete, Ellipsis, Flag, Image, ImagePlay, Images, Italic, Loader2, MapPin, Maximize2, MessageCircleX, MessageSquare, Minus, MinusCircle, Paintbrush, PenLine, PlusCircle, Repeat2, Save, Search, ShieldX, Sliders, Smile, Type, UserRoundPlus, X } from 'lucide-react';
+import { Copy, Delete, Ellipsis, Flag, MessageCircleX, Minus, Repeat2, Save, Search, ShieldX, UserRoundPlus } from 'lucide-react';
 import { Drawer, DrawerClose, DrawerContent, DrawerDescription, DrawerFooter, DrawerHeader, DrawerTitle, DrawerTrigger } from "@/components/ui/drawer";
 import { Button } from '@/components/ui/button';
 import MediaSlide from '../templates/mediaSlides';
@@ -19,16 +19,9 @@ import { updatePost } from '@/redux/postsSlice';
 import { useSocket } from '@/app/providers/SocketProvider';
 import { useUser } from '@/app/providers/UserProvider';
 import { Skeleton } from './ui/skeleton';
-import { DropdownMenu, DropdownMenuContent, DropdownMenuTrigger } from './ui/dropdown-menu';
-import { Select, SelectTrigger, SelectValue, SelectContent, SelectItem } from './ui/select';
-import { Dialog, DialogTrigger, DialogFooter, DialogHeader, DialogContent, DialogTitle, DialogDescription, DialogClose } from './ui/dialog';
-import { cn, formatFileSize, validateFile } from '@/lib/utils';
-import { Input } from './ui/input';
-import { useGlobalFileStorage } from '@/hooks/useFileStorage';
-import { toast } from '@/hooks/use-toast';
-import { FileValidationConfig, TextOverlay } from '@/lib/types/type';
-import CropMediaInterface from './CropMediaInterface';
+import { FileValidationConfig } from '@/lib/types/type';
 import { RootState } from '@/redux/store';
+import ShareButton from './ShareButton';
 
 type PostComponentProps = {
   postData: Post['post'],
@@ -224,9 +217,9 @@ const PostCard = ({ postData, showMedia = true }: PostComponentProps) => {
     <div className='pre-blog' id={data.PostID.slice(0,-4)}>
       <div className='blog !dark:shadow-bar-dark select-none dark:bg-zinc-900 shadow-md dark:text-slate-200' data-id={data.PostID}>
         {postType === 'repost' && (
-          <div className='repost flex items-center gap-2'>
-            <Repeat2 size={20} className='text-brand' />
-            <span className='text-sm'>Reposted by {data?.Username}</span>
+          <div className='repost flex items-center gap-2 text-xs text-brand'>
+            <Repeat2 size={20} />
+            <span>Reposted by {postData?.Username}</span>
           </div>
         )}
         <div className='blogger-details'>
@@ -321,7 +314,7 @@ const PostCard = ({ postData, showMedia = true }: PostComponentProps) => {
             </svg>
             <div className='comments'>{formatNo(data.NoOfComment)}</div>
           </div>
-          <ShareButton post={data} open={open} setOpen={setOpen}>
+          <ShareButton post={data}>
             <svg className={data.Shared ? 'share-icon clicked' : 'share-icon'} width='30px' height='30px' viewBox='-0.5 0 25 25' fill='none' xmlns='http://www.w3.org/2000/svg' onClick={() => handleClick('shared')}>
               <path className='dark:stroke-slate-200' d='M13.47 4.13998C12.74 4.35998 12.28 5.96 12.09 7.91C6.77997 7.91 2 13.4802 2 20.0802C4.19 14.0802 8.99995 12.45 12.14 12.45C12.34 14.21 12.79 15.6202 13.47 15.8202C15.57 16.4302 22 12.4401 22 9.98006C22 7.52006 15.57 3.52998 13.47 4.13998Z' stroke='#242742fd' strokeWidth='1.1' strokeLinecap='round' strokeLinejoin='round'/>
               </svg>
@@ -396,499 +389,6 @@ function Options({options, open, setOpen} : {options: Option[], open: boolean, s
         ))}
       </PopoverContent>
     </Popover>
-    </>
-  )
-}
-
-function ShareButton({children, post, open, setOpen} : {children: ReactNode, post?: PostData, open: boolean, setOpen: React.Dispatch<React.SetStateAction<boolean>>}) {
-  const { userdata } = useUser();
-  const dispatch = useDispatch();
-  const router = useRouter();
-  const socket = useSocket();
-  const { files, clearFiles, setFiles } = useGlobalFileStorage();
-  const [visibility, setVisibility] = useState("everyone");
-  const [errors, setErrors] = useState([''])
-  const [isDrawerOpen, setIsDrawerOpen] = useState(false);
-  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
-  const [isQuoteModalOpen, setIsQuoteModalOpen] = useState(false);
-  const imageInputRef = useRef<HTMLInputElement>(null);
-  const textAreaRef = useRef<HTMLTextAreaElement>(null);
-  const [txtButton, setTxtButton] = useState(false);
-  const [isEditOpen, setIsEditOpen] = useState(false);
-  const [text, setText] = useState('')
-  const [time, setTime] = useState('')
-  const [isPosting, setIsPosting] = useState(false)
-  const textLimit = 400
-
-  useEffect(() => {
-    if(!post) return;
-
-    const interval = setInterval(() => {
-      setTime(updateLiveTime('getlivetime', post.TimeOfPost));
-    }, 1000);
-    return () => clearInterval(interval); // This is important to clear the interval when the component unmounts
-  }, [post?.TimeOfPost]);
-
-  const handleInput = () => {
-    const textArea = textAreaRef.current;
-    if (textArea) {
-      textArea.style.height = 'auto';
-      textArea.style.height = `${textArea.scrollHeight}px`;
-      setTxtButton(textArea.value.trim() !== '');
-    }
-  };
-
-  const options: Option[] = [
-    ...(!post?.Shared ? [
-      {
-        icon: <Repeat2 size={20} />,
-        text: "Repost",
-        onClick: () => {
-          if(!post || !socket) return;
-
-          dispatch(updatePost({ id: post.PostID, updates: { NoOfShares: post.NoOfShares + 1, Shared: true } }));
-          socket.emit('reactToPost(share)', {
-            action: "share",
-            type: "repost",
-            post: post
-          });
-        }
-      }
-    ] : [
-      {
-        icon: <Repeat2 size={20} />,
-        text: "Undo Repost",
-        onClick: () => {
-          if(!post || !socket) return;
-          dispatch(updatePost({ id: post.PostID, updates: { NoOfShares: post.NoOfShares - 1, Shared: false } }));
-
-          socket?.emit('reactToPost(share)', {
-            action: "unshare",
-            post: post
-          });
-        }
-      }
-    ]),
-    {
-      icon: <PenLine size={20} />,
-      text: "Quote",
-      onClick: () => {
-        // Close other components before opening the quote modal
-        setIsDrawerOpen(false);
-        setIsPopoverOpen(false);
-        setIsQuoteModalOpen(true);
-      }
-    }
-  ]
-
-  // Handle mobile drawer
-  const handleDrawerChange = (open: boolean) => {
-    setIsDrawerOpen(open);
-    setOpen(open);
-  };
-
-  // Handle desktop popover
-  const handlePopoverChange = (open: boolean) => {
-    setIsPopoverOpen(open);
-    setOpen(open);
-  };
-
-  function handleFiles(e: ChangeEvent<HTMLInputElement>): void {
-    if (!imageInputRef.current) return;
-
-    const mediaFiles = e.target.files;
-    if (!mediaFiles) return;
-
-    const newFiles = Array.from(mediaFiles);
-    const validationErrors: string[] = [];
-    const validFiles: File[] = [];
-
-    // Check total number of files
-    if (files.length > FILE_VALIDATION_CONFIG.maxFiles) {
-      toast({
-        title: 'Warning',
-        description: `Maximum ${FILE_VALIDATION_CONFIG.maxFiles} files allowed`,
-        variant: 'destructive'
-      });
-      clearFiles();
-      return;
-    }
-
-    // Calculate total size including existing files
-    const existingSize = files.reduce((acc, file) => acc + file.size, 0);
-    const newTotalSize = newFiles.reduce((acc, file) => acc + file.size, existingSize);
-
-    if (newTotalSize > FILE_VALIDATION_CONFIG.maxTotalSize) {
-      toast({
-        title: 'Warning',
-        description: `Total size exceeds ${formatFileSize(FILE_VALIDATION_CONFIG.maxTotalSize)}`,
-        variant: 'destructive'
-      });
-      return;
-    }
-
-    // Validate each file
-    newFiles.forEach(file => {
-      try {
-        validateFile(file, FILE_VALIDATION_CONFIG);
-        validFiles.push(file);
-      } catch (error) {
-        if (error instanceof Error) {
-          validationErrors.push(error.message);
-        }
-      }
-    });
-
-    // Update state
-    if (validationErrors.length > 0) {
-      setErrors(validationErrors);
-      validationErrors.forEach(error => toast({
-        title: 'Warning',
-        description: error
-      }));
-    }
-
-    if (validFiles.length > 0) {
-      setFiles(validFiles);
-    }
-
-    // Reset input
-    e.target.value = '';
-  }
-
-
-  const handlePost = async () => {
-    if(!socket || !post) return;
-    setIsPosting(true)
-    
-    try {
-      let media: string[] = []
-      if(files.length > 0) {
-
-        files.forEach(async (file) => {
-          const response = await fetch(
-            "https://localhost:3000" + '/api/upload',
-            {
-              method: 'POST',
-              headers: {
-                'Content-Type': 'application/json',
-              },
-              body: JSON.stringify({ filename: file!.name, contentType: file!.type, bucketName: 'post-s' }),
-            }
-          );
-  
-          if (!response.ok) {
-            throw new Error('Failed to upload image');
-          }
-  
-          const { url, fields } = await response.json();
-          const formData1 = new FormData();
-  
-          for (const key in fields) {
-            formData1.append(key, fields[key]);
-          }
-  
-          formData1.append('file', file);
-  
-          const uploadResponse = await fetch(url, {
-            method: 'POST',
-            body: formData1,
-          });
-  
-          if (!uploadResponse.ok) { 
-            throw new Error('Failed to upload image');
-          } else {
-            // formData.file = url+fields.key;
-            media.push(url+fields.key);
-          }
-        })
-
-      }
-
-      const quote: Partial<PostData> = {
-        Visibility: visibility as PostData['Visibility'],
-        Caption: text,
-        Image: media,
-        Code: '',
-        WhoCanComment: visibility as PostData['WhoCanComment'],
-        OriginalPostId: post.PostID
-      }
-
-      socket.emit('reactToPost(share)', {
-        action: "share",
-        type: "quote",
-        post: quote
-      });
-
-    } catch (error: any) {
-      console.log(error)
-    } finally {
-      setFiles([]);
-      setIsQuoteModalOpen(false);
-      setIsPosting(false)
-    }
-  };
-
-  if (!userdata._id) return <div className='flex items-center justify-center gap-2 cursor-not-allowed'>{children}</div>;
-
-  return (
-    <>
-      {/* Mobile view */}
-      <Drawer open={isDrawerOpen} onOpenChange={handleDrawerChange}>
-        <DrawerTrigger className='tablets:hidden flex items-center justify-center gap-2 cursor-pointer'>
-          {children}
-        </DrawerTrigger>
-        <DrawerContent className='tablets:hidden'>
-          <DrawerHeader className="text-left">
-            <DrawerTitle className='hidden'>Options</DrawerTitle>
-            <DrawerDescription className='flex flex-col gap-2'>
-              {options.map(({ icon, text, onClick }, index) => {
-                if(text === "Quote") return (
-                  <span 
-                    key={index} 
-                    className='flex gap-1 items-center cursor-pointer w-full' 
-                    onClick={() => {
-                      setIsDrawerOpen(false);
-                      setIsQuoteModalOpen(true);
-                    }}
-                  >
-                    {icon}
-                    <span className='dark:text-white'>{text}</span>
-                  </span>
-                )
-                return (
-                  <span key={index} className='flex gap-1 items-center cursor-pointer w-full' onClick={onClick}>
-                    {icon}
-                    <span className='dark:text-white'>{text}</span>
-                  </span>
-                )
-              })}
-            </DrawerDescription>
-          </DrawerHeader>
-          <DrawerFooter className="pt-2">
-            <DrawerClose asChild>
-              <Button variant="outline">Cancel</Button>
-            </DrawerClose>
-          </DrawerFooter>
-        </DrawerContent>
-      </Drawer>
-
-      {/* Desktop view */}
-      <DropdownMenu open={isPopoverOpen} onOpenChange={handlePopoverChange}>
-        <DropdownMenuTrigger className='hidden tablets:flex items-center justify-center gap-2 cursor-pointer'>
-          {children}
-        </DropdownMenuTrigger>
-        <DropdownMenuContent className='bg-white hidden tablets:block dark:bg-zinc-800 w-auto space-y-2 mt-1 mr-2 p-2 rounded-md shadow-lg'>
-          {options.map(({ icon, text, onClick }, index) => {
-            if(text === "Quote") return (
-              <Button 
-                key={index} 
-                className='flex gap-1 items-center cursor-pointer w-full bg-white dark:bg-zinc-900 text-black dark:text-white hover:bg-white/90 dark:hover:bg-zinc-900/90' 
-                onClick={() => {
-                  setIsPopoverOpen(false);
-                  setIsQuoteModalOpen(true);
-                }}
-              >
-                {icon}
-                <span className='dark:text-white'>{text}</span>
-              </Button>
-            )
-            return (
-              <Button key={index} className='flex gap-1 items-center cursor-pointer w-full bg-white dark:bg-zinc-900 text-black dark:text-white hover:bg-white/90 dark:hover:bg-zinc-900/90' onClick={onClick}>
-                {icon}
-                <span className='dark:text-white'>{text}</span>
-              </Button>
-            )
-          })}
-        </DropdownMenuContent>
-      </DropdownMenu>
-
-      {/* Separate Dialog component */}
-      <Dialog open={isQuoteModalOpen} onOpenChange={setIsQuoteModalOpen}>
-        <DialogContent className="bg-white dark:bg-zinc-900 overflow-auto mb:h-full">
-          
-          <DialogHeader className='dark:text-white'>
-            <DialogTitle className='text-center'></DialogTitle>
-            <DialogDescription>
-              <span className="flex justify-between items-center">
-                <DialogClose className="dark:text-white hover:bg-gray-800 p-2 rounded-full transition-all duration-200 transform hover:scale-110">
-                  <X size={16} />
-                </DialogClose>
-                <span className="text-brand text-xl font-bold">Drafts</span>
-              </span>
-            </DialogDescription>
-          </DialogHeader>
-          <div className="max-h-80 overflow-auto">
-
-            {/* User Info and Dropdown */}
-            <div className="flex items-center space-x-3">
-              <div className="w-10 h-10 rounded-full overflow-hidden">
-                {userdata.dp ?
-                <img
-                  src={userdata.dp}
-                  alt="Profile"
-                  className="w-full h-full object-cover"
-                /> :
-                <Skeleton className="w-full h-full object-cover"/>}
-              </div>
-              <Select onValueChange={setVisibility} defaultValue="everyone" value={visibility}>
-                <SelectTrigger className="max-w-[120px] rounded-2xl">
-                  <SelectValue placeholder="Everyone" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="everyone">Everyone</SelectItem>
-                  <SelectItem value="none">None</SelectItem>
-                  <SelectItem value="friends">Friends</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-
-            <Input
-              accept=".jpg, .jpeg, .png, .gif"
-              id="images"
-              ref={imageInputRef} 
-              type="file"
-              multiple 
-              onChange={handleFiles} 
-              className="hidden"
-            />
-
-            {/* Comment Input */}
-            <textarea
-              ref={textAreaRef} 
-              value={text}
-              maxLength={textLimit}
-              spellCheck
-              onChange={(e) => setText(e.target.value.substring(0, textLimit))}
-              placeholder="Add a comment" 
-              className="flex-grow text-sm w-full mt-2 max-h-[400px] focus:border-b focus:border-brand resize-none bg-transparent border-none focus:outline-none dark:text-white"
-              onInput={handleInput}
-            />
-
-            {/** Media */}
-            {files.length > 0 && (
-              <div className="grid grid-cols-2 gap-2 my-2">
-                {files.map((file, index) => (
-                  <div key={index} className="relative group">
-                    {file.type.startsWith('image/') ? (
-                      <img
-                        src={URL.createObjectURL(file)}
-                        alt={file.name}
-                        className="w-full h-40 object-cover rounded-lg"
-                      />
-                    ) : file.type.startsWith('video/') ? (
-                      <video
-                        src={URL.createObjectURL(file)}
-                        className="w-full h-40 object-cover rounded-lg"
-                        controls
-                      />
-                    ) : null}
-                    <CropMediaInterface files={files} setFiles={setFiles} imageIndex={index}>
-                      <button
-                        onClick={() => {
-                          setIsEditOpen(true)
-                        }}
-                        className="absolute top-2 left-2 p-1 rounded-full bg-black/50 text-white opacity-0 group-hover:opacity-100 transition-opacity"
-                      >
-                        <Paintbrush size={16} />
-                      </button>
-                    </CropMediaInterface>
-                    <button
-                      onClick={() => {
-                        const newFiles = files.filter((_, i) => i !== index);
-                        setFiles(newFiles);
-                      }}
-                      className="absolute top-2 right-2 p-1 rounded-full bg-black/50 text-white opacity-0 group-hover:opacity-100 transition-opacity"
-                    >
-                      <X size={16} />
-                    </button>
-                  </div>
-                ))}
-              </div>
-            )}
-
-            {/* Post Card */}
-            {post && (
-              <div className="border border-gray-800 rounded-xl p-4">
-                <div className="flex items-start space-x-3">
-                  <div className="w-10 h-10 rounded-full overflow-hidden flex-shrink-0">
-                    <img
-                      src={post.DisplayPicture}
-                      alt={post.Username}
-                      className="w-full h-full object-cover"
-                    />
-                  </div>
-                  <div className="flex-1 grid gap-2 grid-cols-2">
-                    <div className="flex items-center col-span-2 flex-wrap">
-                      <span className="font-bold dark:text-white mr-1 truncate">{post.NameOfPoster}</span>
-                      {post.Verified && (
-                        <svg className="w-4 h-4 text-brand fill-current" viewBox="0 0 24 24">
-                          <path d="M22.5 12.5c0-1.58-.875-2.95-2.148-3.6.154-.435.238-.905.238-1.4 0-2.21-1.71-3.998-3.818-3.998-.47 0-.92.084-1.336.25C14.818 2.415 13.51 1.5 12 1.5s-2.816.917-3.437 2.25c-.415-.165-.866-.25-1.336-.25-2.11 0-3.818 1.79-3.818 4 0 .494.083.964.237 1.4-1.272.65-2.147 2.018-2.147 3.6 0 1.495.782 2.798 1.942 3.486-.02.17-.032.34-.032.514 0 2.21 1.708 4 3.818 4 .47 0 .92-.086 1.335-.25.62 1.334 1.926 2.25 3.437 2.25 1.512 0 2.818-.916 3.437-2.25.415.163.865.248 1.336.248 2.11 0 3.818-1.79 3.818-4 0-.174-.012-.344-.033-.513 1.158-.687 1.943-1.99 1.943-3.484zm-6.616-3.334l-4.334 6.5c-.145.217-.382.334-.625.334-.143 0-.288-.04-.416-.126l-.115-.094-2.415-2.415c-.293-.293-.293-.768 0-1.06s.768-.294 1.06 0l1.77 1.767 3.825-5.74c.23-.345.696-.436 1.04-.207.346.23.44.696.21 1.04z" />
-                      </svg>
-                      )}
-                      <span className="text-gray-500 ml-1">@{post.Username} · {time}</span>
-                    </div>
-                    {/* <div className="text-gray-500 text-sm mb-2">Replying to @NintendoAmerica</div> */}
-                    {post.Caption ? (
-                      <p className={`dark:text-white text-sm mb-2 ${post.Image.length > 0 ? '' : 'col-span-2'} whitespace-pre-wrap`}>{post.Caption.length > 250 ? post.Caption.substring(0, 250) + '...' : post.Caption}</p>
-                    ) : null}
-                    {/* {showMore} */}
-                    {(post.Image.length > 0) &&
-                      <MediaSlide className={`overflow-auto rounded-lg ${!post.Caption.length ? 'col-span-2' : ''}`} postData={post}/>
-                    }
-                  </div>
-                </div>
-              </div>
-            )}
-          </div>
-          
-          <div className='sticky bottom-0 space-y-2'>
-            {/* Everyone Can Reply */}
-            <div className="bg-white dark:bg-zinc-900 flex items-center text-brand space-x-2">
-              <CircleAlert size={16} />
-              <span>{visibility[0].toUpperCase() + visibility.slice(1)} can reply</span>
-            </div>
-
-            {/* Character Counter */}
-            <div className={`flex justify-end text-gray-500 text-sm mb-2 ${text.length === textLimit ? 'text-red-500' : ''}`}>
-              <span>{text.length === textLimit && 'You have reached the text characters limit! • '}{text.length}/{textLimit}</span>
-            </div>
-
-            <DialogFooter className='bg-white dark:bg-zinc-900'>
-              {/* Bottom Toolbar */}
-              <div className="border-t border-gray-800 flex-1 py-3">
-                <div className="gap-8 grid grid-cols-2 mb:grid-cols-1">
-                  <div className="flex justify-between items-center">
-                    <button onClick={() => imageInputRef?.current?.click()} className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                      <Images size={16} className='group-hover:text-white'/>
-                    </button>
-                    <button className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                      <CircleCheck size={16} className='group-hover:text-white'/>
-                    </button>
-                    <button className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                      <ChartBarDecreasing size={16} className='group-hover:text-white'/>
-                    </button>
-                    <button className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                      <Smile size={16} className='group-hover:text-white'/>
-                    </button>
-                    <button className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                      <Clock4 size={16} className='group-hover:text-white'/>
-                    </button>
-                    <button className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                      <MapPin size={16} className='group-hover:text-white'/>
-                    </button>
-                  </div>
-
-                  <Button disabled={!txtButton || text.length === textLimit || !text.length} onClick={handlePost} className="bg-brand hover:bg-brand/60 text-white w-full my-2 tablets:w-auto tablets:my-0 font-bold py-2 px-6 rounded-full transition-all duration-200 transform hover:scale-105">
-                    {isPosting ? 'Posting...' : 'Post'}
-                    {isPosting && <Loader2 className='animate-spin ml-2' size={16} />}
-                  </Button>
-                </div>
-              </div>
-            </DialogFooter>
-          </div>
-        </DialogContent>
-      </Dialog> 
     </>
   )
 }
